Rename implementations of concrete RelativeTimeDistortions
