+++ src/ConstantTimeCompression.ts
