+++ src/ConstantTimeDilation.ts
